refactor render
