+++ src/nerte/render/projection_unittest.py
@@ -13,13 +13,14 @@ from nerte.base_test_case import BaseTestCase
 
 from nerte.values.coordinates import Coordinates3D, Coordinates2D
 from nerte.values.coordinates_unittest import coordinates_2d_equiv
-from nerte.values.interval import Interval
 from nerte.values.linalg import AbstractVector
 from nerte.values.tangential_vector import TangentialVector
-from nerte.values.manifolds.cartesian import Plane
+from nerte.values.interval import Interval
+from nerte.values.domains import CartesianProduct2D
+from nerte.values.charts.cartesian import Plane
 from nerte.world.camera import Camera
-from nerte.geometry.cartesian_geometry import CartesianGeometry
-from nerte.geometry.cartesian_geometry_unittest import cartesian_ray_equiv
+from nerte.geometry.geometry import StandardGeometry
+from nerte.geometry.geometry_unittest import standard_ray_equiv
 from nerte.render.projection import (
     detector_manifold_coords,
     orthographic_ray_for_pixel,
@@ -34,17 +35,17 @@ class DetectorManifoldCoordsTest(BaseTestCase):
     def setUp(self) -> None:
         # camera
         loc = Coordinates3D((0.0, 0.0, 0.0))
-        domain = Interval(-1.0, 1.0)
+        interval = Interval(-1.0, 1.0)
+        domain = CartesianProduct2D(interval, interval)
         manifold = Plane(
             AbstractVector((1.0, 0.0, 0.0)),
             AbstractVector((0.0, 1.0, 0.0)),
-            x0_domain=domain,
-            x1_domain=domain,
             offset=AbstractVector((0.0, 0.0, 1.0)),
         )
         dim = 10
         self.camera = Camera(
             location=loc,
+            detector_domain=domain,
             detector_manifold=manifold,
             canvas_dimensions=(dim, dim),
         )
@@ -101,22 +102,22 @@ class OrthographicProjectionTest(BaseTestCase):
     def setUp(self) -> None:
         # camera
         loc = Coordinates3D((0.0, 0.0, 0.0))
-        domain = Interval(-1.0, 1.0)
+        interval = Interval(-1.0, 1.0)
+        domain = CartesianProduct2D(interval, interval)
         manifold = Plane(
             AbstractVector((1.0, 0.0, 0.0)),
             AbstractVector((0.0, 1.0, 0.0)),
-            x0_domain=domain,
-            x1_domain=domain,
             offset=AbstractVector((0.0, 0.0, 1.0)),
         )
         dim = 100
         self.camera = Camera(
             location=loc,
+            detector_domain=domain,
             detector_manifold=manifold,
             canvas_dimensions=(dim, dim),
         )
         # geometry
-        self.geometry = CartesianGeometry()
+        self.geometry = StandardGeometry()
         # pixels and rays
         self.pixel_locations = (
             (0, 0),
@@ -133,7 +134,7 @@ class OrthographicProjectionTest(BaseTestCase):
             (dim + 1, 1),
         )
         # orthographic ray from detecor manifold coordinates
-        def make_ray(coords2d: Coordinates2D) -> CartesianGeometry.Ray:
+        def make_ray(coords2d: Coordinates2D) -> StandardGeometry.Ray:
             return self.geometry.ray_from_tangent(
                 TangentialVector(
                     point=manifold.embed(coords2d),
@@ -159,9 +160,9 @@ class OrthographicProjectionTest(BaseTestCase):
                 geometry=self.geometry,
                 pixel_location=pix_loc,
             )
-            self.assertIsInstance(ray, CartesianGeometry.Ray)
-            cart_ray = cast(CartesianGeometry.Ray, ray)
-            self.assertPredicate2(cartesian_ray_equiv, cart_ray, pix_ray)
+            self.assertIsInstance(ray, StandardGeometry.Ray)
+            cart_ray = cast(StandardGeometry.Ray, ray)
+            self.assertPredicate2(standard_ray_equiv, cart_ray, pix_ray)
 
     def test_orthographic_ray_for_pixel_invalid_values(self) -> None:
         """Tests orthographic projection for pixel's invalid values."""
@@ -182,22 +183,22 @@ class PerspectiveProjectionTest(BaseTestCase):
     def setUp(self) -> None:
         # camera
         loc = Coordinates3D((0.0, 0.0, 0.0))
-        domain = Interval(-1.0, 1.0)
+        interval = Interval(-1.0, 1.0)
+        domain = CartesianProduct2D(interval, interval)
         manifold = Plane(
             AbstractVector((1.0, 0.0, 0.0)),
             AbstractVector((0.0, 1.0, 0.0)),
-            x0_domain=domain,
-            x1_domain=domain,
             offset=AbstractVector((0.0, 0.0, 1.0)),
         )
         dim = 100
         self.camera = Camera(
             location=loc,
+            detector_domain=domain,
             detector_manifold=manifold,
             canvas_dimensions=(dim, dim),
         )
         # geometry
-        self.geometry = CartesianGeometry()
+        self.geometry = StandardGeometry()
         # pixels and rays
         self.pixel_locations = (
             (0, 0),
@@ -216,7 +217,7 @@ class PerspectiveProjectionTest(BaseTestCase):
         # perspective ray from detecor manifold coordinates
         def make_ray(
             coords2d: Coordinates2D,
-        ) -> CartesianGeometry.Ray:
+        ) -> StandardGeometry.Ray:
             return self.geometry.ray_from_coords(
                 start=self.camera.location,
                 target=self.camera.detector_manifold.embed(coords2d),
@@ -240,9 +241,9 @@ class PerspectiveProjectionTest(BaseTestCase):
                 geometry=self.geometry,
                 pixel_location=pix_loc,
             )
-            self.assertIsInstance(ray, CartesianGeometry.Ray)
-            cart_ray = cast(CartesianGeometry.Ray, ray)
-            self.assertPredicate2(cartesian_ray_equiv, cart_ray, pix_ray)
+            self.assertIsInstance(ray, StandardGeometry.Ray)
+            cart_ray = cast(StandardGeometry.Ray, ray)
+            self.assertPredicate2(standard_ray_equiv, cart_ray, pix_ray)
 
     def test_perspective_ray_for_pixel_invalid_values(self) -> None:
         """Tests perspective projection for pixel's invalid values."""
@@ -263,22 +264,22 @@ class ObscuraProjectionTest(BaseTestCase):
     def setUp(self) -> None:
         # camera
         loc = Coordinates3D((0.0, 0.0, 1.0))
-        domain = Interval(-1.0, 1.0)
+        interval = Interval(-1.0, 1.0)
+        domain = CartesianProduct2D(interval, interval)
         manifold = Plane(
             AbstractVector((1.0, 0.0, 0.0)),
             AbstractVector((0.0, 1.0, 0.0)),
-            x0_domain=domain,
-            x1_domain=domain,
             offset=AbstractVector((0.0, 0.0, 0.0)),
         )
         dim = 100
         self.camera = Camera(
             location=loc,
+            detector_domain=domain,
             detector_manifold=manifold,
             canvas_dimensions=(dim, dim),
         )
         # geometry
-        self.geometry = CartesianGeometry()
+        self.geometry = StandardGeometry()
         # pixels and rays
         self.pixel_locations = (
             (0, 0),
@@ -297,7 +298,7 @@ class ObscuraProjectionTest(BaseTestCase):
         # perspective ray from detecor manifold coordinates
         def make_ray(
             coords2d: Coordinates2D,
-        ) -> CartesianGeometry.Ray:
+        ) -> StandardGeometry.Ray:
             return self.geometry.ray_from_coords(
                 start=self.camera.detector_manifold.embed(coords2d),
                 target=self.camera.location,
@@ -321,9 +322,9 @@ class ObscuraProjectionTest(BaseTestCase):
                 geometry=self.geometry,
                 pixel_location=pix_loc,
             )
-            self.assertIsInstance(ray, CartesianGeometry.Ray)
-            cart_ray = cast(CartesianGeometry.Ray, ray)
-            self.assertPredicate2(cartesian_ray_equiv, cart_ray, pix_ray)
+            self.assertIsInstance(ray, StandardGeometry.Ray)
+            cart_ray = cast(StandardGeometry.Ray, ray)
+            self.assertPredicate2(standard_ray_equiv, cart_ray, pix_ray)
 
     def test_perspective_ray_for_pixel_invalid_values(self) -> None:
         """Tests perspective projection for pixel's invalid values."""
