+++ src/nerte/render/renderer_unittest.py
@@ -9,15 +9,15 @@ import unittest
 from nerte.base_test_case import BaseTestCase
 
 from nerte.values.coordinates import Coordinates3D
-from nerte.values.interval import Interval
 from nerte.values.linalg import AbstractVector
-from nerte.values.manifolds.cartesian import Plane
+from nerte.values.interval import Interval
+from nerte.values.domains import CartesianProduct2D
+from nerte.values.charts.cartesian import Plane
 from nerte.values.face import Face
 from nerte.world.object import Object
 from nerte.world.camera import Camera
 from nerte.world.scene import Scene
-from nerte.geometry.geometry import Geometry
-from nerte.geometry.cartesian_geometry import CartesianGeometry
+from nerte.geometry.geometry import Geometry, StandardGeometry
 from nerte.render.renderer import Renderer
 
 
@@ -34,15 +34,17 @@ class RendererTest(BaseTestCase):
         # camera
         loc = Coordinates3D((0.0, 0.0, -10.0))
         interval = Interval(-1.0, 1.0)
+        domain = CartesianProduct2D(interval, interval)
         manifold = Plane(
             AbstractVector((1.0, 0.0, 0.0)),
             AbstractVector((0.0, 1.0, 0.0)),
-            x0_domain=interval,
-            x1_domain=interval,
+            interval0=interval,
+            interval1=interval,
         )
         dim = 25
         cam = Camera(
             location=loc,
+            detector_domain=domain,
             detector_manifold=manifold,
             canvas_dimensions=(dim, dim),
         )
@@ -50,7 +52,7 @@ class RendererTest(BaseTestCase):
         self.scene = Scene(camera=cam)
         self.scene.add_object(obj)
         # geometry
-        self.geometry = CartesianGeometry()
+        self.geometry = StandardGeometry()
 
     def test_render_implementation(self) -> None:
         """Tests Render implementation."""
